+++ src/main/java/com/sprint/mople/domain/dm/repository/FollowRepository.java
@@ -1,6 +1,6 @@
-package com.sprint.mople.dm.repository;
+package com.sprint.mople.domain.dm.repository;
 
-import com.sprint.mople.follow.entity.Follow;
+import com.sprint.mople.domain.follow.entity.Follow;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
